Got some basic BBCode support in a really janky way!
So basically the default tooltip node for the editor inspector is a kind of stripped down RichTextLabel. I need to do some research to figure out exactly what's happening, but just by external probing, I was able to insert terminating tags for the bold and underline that is forced for the tooltip, then I can just append any standard BBCode tooltip entries I like.
Somewhere in the internals, the terminating bold and underline are appended to the tooltip, which leaves a weird tag at the end, but I just put a few new lines before it to make it very separate from the rest of the text.
Realistically, it works. It's hella jank, but it works and I can have FORMATTED TOOLTIPS FOR CUSTOM EXPORT VARS.
Like holy hell that is being added in Godot 4 and I just fucking made it in a day